refactor: change layout of OTPText
