+++ petJournal/app/src/main/java/com/soujunior/petjournal/ui/screens_app/account_manager/awaitingCodeScreen/components/VerificationCodeInput.kt
@@ -39,7 +39,8 @@ fun VerificationCodeInput(viewModel: AwaitingCodeViewModel) {
                 )
             )
         },
-        textError = state.codeOTPError
+        textError = state.codeOTPError,
+        viewModel = viewModel
     )
     Box(
         modifier = Modifier
